refactor with int32 indices, string_view parsing. move fuse flag in exenetwork.

+++ onnxruntime/core/providers/openvino/backend_manager.cc
@@ -781,7 +781,7 @@ void BackendManager::RewindKVCache(size_t index) {
   }
 }
 
-void BackendManager::ReorderKVCache(const std::vector<size_t>& src_indices, const std::vector<size_t>& dst_indices) {
+void BackendManager::ReorderKVCache(const std::vector<int32_t>& src_indices, const std::vector<int32_t>& dst_indices) {
   if (concrete_backend_) {
     concrete_backend_->ReorderKVCache(src_indices, dst_indices);
   }
